Move compiled function classes into new module
